Organização em packages proposta
simplesmente organizei diferente a tarefa 4 que estava no git
