--- tarefa4/src/classes/interfaces/AcaoDeCombate.java
@@ -1,5 +1,0 @@
-package classes.interfaces;
-
-public interface AcaoDeCombate {
-    public void executar(Combatente usuario, Combatente alvo);
-}
