--- tarefa4/src/classes/interfaces/Combatente.java
@@ -1,16 +1,0 @@
-package classes.interfaces;
-
-import classes.armas.Arma;
-
-public interface Combatente {
-    public String getNome();
-    public int receberDano(int dano);
-    public void receberCura(int cura);
-    public AcaoDeCombate escolherAcao();
-    public int getPontosDeVida();
-    public int getForca();
-    public float getSorte();
-    public Arma getArma();
-    public int getNivel();
-    public void darDano(Combatente alvo, int dano);
-}
