--- tarefa4/src/classes/interfaces/Evento.java
@@ -1,6 +1,0 @@
-package classes.interfaces;
-
-public interface Evento {
-    public boolean verificarGatilho(); // TODO: não faz sentido ser publico. tem que ser privado e chamado automaticamente pelo metodo executar().
-    public void executar();
-}
