--- tarefa4/src/classes/interfaces/GeradorDeFases.java
@@ -1,7 +1,0 @@
-package classes.interfaces;
-
-import java.util.List;
-
-public interface GeradorDeFases {
-    public List<Fase> gerar(int quantidadeDeFases);
-}
