--- tarefa4/src/classes/interfaces/Item.java
@@ -1,5 +1,0 @@
-package classes.interfaces;
-
-public interface Item {
-    public String getNome();
-}
