--- tarefa4/src/classes/interfaces/Lootavel.java
@@ -1,5 +1,0 @@
-package classes.interfaces;
-
-public interface Lootavel extends Combatente{
-    public Item droparLoot();
-}
