+++ tarefa4/src/com/rpglab/cenario/Fase.java
@@ -1,0 +1,11 @@
+package com.rpglab.cenario;
+
+import com.rpglab.combate.Evento;
+import com.rpglab.personagens.heroi.Herói;
+
+public interface Fase {
+    public boolean iniciar(Herói heroi);
+    public boolean isConcluida();
+    public TipoCenario getTipoCenario();
+    public void adicionarEvento(Evento evento);
+}
