+++ tarefa4/src/com/rpglab/cenario/GeradorDeFases.java
@@ -1,0 +1,7 @@
+package com.rpglab.cenario;
+
+import java.util.List;
+
+public interface GeradorDeFases {
+    public List<Fase> gerar(int quantidadeDeFases);
+}
