+++ tarefa4/src/com/rpglab/combate/AcaoDeCombate.java
@@ -1,0 +1,7 @@
+package com.rpglab.combate;
+
+import com.rpglab.personagens.Combatente;
+
+public interface AcaoDeCombate {
+    public void executar(Combatente usuario, Combatente alvo);
+}
