+++ tarefa4/src/com/rpglab/combate/Ataque.java
@@ -1,0 +1,20 @@
+package com.rpglab.combate;
+
+import com.rpglab.itens.Arma;
+import com.rpglab.personagens.Combatente;
+
+// Ataque padrão, usado por todods
+public class Ataque implements AcaoDeCombate{
+    public Ataque(){}
+
+    public void executar(Combatente usuario, Combatente alvo){
+        Arma arma = usuario.getArma();
+        float dano = usuario.getForca() * usuario.getSorte() * 2 + usuario.getNivel() * 0.5f + arma.getDano();
+        if (arma == Arma.DESARMADO)
+            System.out.println(usuario.getNome() + " ataca desarmado.");
+        else 
+            System.out.println(usuario.getNome() + " ataca com " + arma.getNome() + ".");
+        usuario.darDano(alvo, Math.round(dano));
+        
+    }
+}
