+++ tarefa4/src/com/rpglab/combate/Evento.java
@@ -1,0 +1,6 @@
+package com.rpglab.combate;
+
+public interface Evento {
+    public boolean verificarGatilho(); // TODO: não faz sentido ser publico. tem que ser privado e chamado automaticamente pelo metodo executar().
+    public void executar();
+}
