+++ tarefa4/src/com/rpglab/itens/Item.java
@@ -1,0 +1,5 @@
+package com.rpglab.itens;
+
+public interface Item {
+    public String getNome();
+}
