+++ tarefa4/src/com/rpglab/personagens/Combatente.java
@@ -1,0 +1,17 @@
+package com.rpglab.personagens;
+
+import com.rpglab.combate.AcaoDeCombate;
+import com.rpglab.itens.Arma;
+
+public interface Combatente {
+    public String getNome();
+    public int receberDano(int dano);
+    public void receberCura(int cura);
+    public AcaoDeCombate escolherAcao();
+    public int getPontosDeVida();
+    public int getForca();
+    public float getSorte();
+    public Arma getArma();
+    public int getNivel();
+    public void darDano(Combatente alvo, int dano);
+}
