+++ tarefa4/src/com/rpglab/personagens/Lootavel.java
@@ -1,0 +1,7 @@
+package com.rpglab.personagens;
+
+import com.rpglab.itens.Item;
+
+public interface Lootavel extends Combatente{
+    public Item droparLoot();
+}
